requirements.txt
discord.py
ics==0.7.1
schedule==1.1.0
apscheduler==3.9.1
requests==2.26.0
PyGithub